requirements.txt
pynput==1.7.6
Pillow==10.2.0
requests==2.31.0
platformdirs==4.2.0
psutil==5.9.8
sounddevice==0.4.6
geocoder==1.38.1
scipy==1.12.0